LayoutTests/imported/w3c/IndexedDB-private-browsing/idbobjectstore_put12-expected.txt


PASS IDBObjectStore.put() - Attempt to put a record where the record's in-line key is not defined
